requirements.txt
Django==2.2
Pillow==7.2.0
psycopg2==2.7.7
psycopg2-binary==2.8.6
pytz==2020.1
sorl-thumbnail==12.6.3
sqlparse==0.3.1

djangorestframework~=3.12.2